fix: loading lottie preview json loads correctly
