+++ apps/lottie-viewer/src/locations/Field.test.tsx
@@ -1,52 +1,26 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { render, screen, fireEvent } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import { useSDK } from '@contentful/react-apps-toolkit';
 import Field from './Field';
 
-// Mocks
 vi.mock('@contentful/react-apps-toolkit', () => ({
   useSDK: vi.fn(),
 }));
 
-vi.mock('@lottiefiles/dotlottie-react', () => ({
-  DotLottieReact: () => <div data-testid="lottie-player">Lottie Player</div>,
-}));
-
-vi.mock('@contentful/field-editor-json', () => ({
-  JsonEditor: () => <div data-testid="json-editor">JSON Editor</div>,
-}));
-
-vi.mock('@monaco-editor/react', () => ({
+vi.mock('@src/components/field/LottiePreviewField', () => ({
   __esModule: true,
-  default: ({ onMount }: any) => {
-    const mockEditor = {
-      getModel: () => ({
-        getFullModelRange: () => ({ startLineNumber: 1, endLineNumber: 1 }),
-        getValue: () => '{}',
-        setValue: vi.fn(),
-        pushStackElement: vi.fn(),
-        canUndo: () => true,
-        canRedo: () => true,
-      }),
-      pushUndoStop: vi.fn(),
-      executeEdits: vi.fn(),
-      trigger: vi.fn(),
-      onDidChangeModelContent: vi.fn(),
-    };
-    onMount?.(mockEditor);
-    return <div data-testid="monaco-editor">Editor</div>;
-  },
+  default: ({ lottieJson }: { lottieJson: any }) => (
+    <div data-testid="lottie-preview-field">{JSON.stringify(lottieJson)}</div>
+  ),
 }));
 
-describe('Field Component', () => {
+describe('Field', () => {
   const mockSdk = {
     window: {
       startAutoResizer: vi.fn(),
     },
     field: {
       getValue: vi.fn(),
-      onValueChanged: vi.fn(),
-      setValue: vi.fn(),
     },
   };
 
@@ -55,52 +29,24 @@ describe('Field Component', () => {
     (useSDK as any).mockReturnValue(mockSdk);
   });
 
-  it('renders without crashing', () => {
-    render(<Field />);
-    expect(screen.getByText('Undo')).toBeTruthy();
-    expect(screen.getByText('Redo')).toBeTruthy();
-    expect(screen.getByTestId('monaco-editor')).toBeTruthy();
-  });
-
-  it('initializes with correct SDK setup', () => {
+  it('calls sdk methods on mount', () => {
+    mockSdk.field.getValue.mockReturnValue({ v: '5.5.7', fr: 30 });
     render(<Field />);
     expect(mockSdk.window.startAutoResizer).toHaveBeenCalled();
     expect(mockSdk.field.getValue).toHaveBeenCalled();
-    expect(mockSdk.field.onValueChanged).toHaveBeenCalled();
   });
 
-  it('renders Lottie player when JSON data is available', () => {
-    mockSdk.field.getValue.mockReturnValue({ some: 'data' });
+  it('renders LottiePreviewField with value from SDK', () => {
+    const testJson = { v: '5.5.7', fr: 24, ip: 0, op: 60 };
+    mockSdk.field.getValue.mockReturnValue(testJson);
+
     render(<Field />);
-    expect(screen.getByTestId('lottie-player')).toBeTruthy();
+    expect(screen.getByTestId('lottie-preview-field')).toBeTruthy();
   });
 
-  it('still renders Lottie player when no JSON is available', () => {
+  it('renders LottiePreviewField with empty object if getValue is null', () => {
     mockSdk.field.getValue.mockReturnValue(null);
     render(<Field />);
-    expect(screen.getByTestId('lottie-player')).toBeTruthy();
-  });
-
-  it('clears JSON when Clear is clicked', () => {
-    render(<Field />);
-    const clearButton = screen.getByText('Clear');
-    fireEvent.click(clearButton);
-    expect(mockSdk.field.setValue).toHaveBeenCalledWith({});
-  });
-
-  it('opens and closes the JSON modal', () => {
-    render(<Field />);
-    const modalOpen = screen.getByLabelText('preview-lottie-json');
-    fireEvent.click(modalOpen);
-    expect(screen.getByText('Lottie Preview - JSON editor')).toBeTruthy();
-  });
-
-  it('handles undo and redo button clicks', () => {
-    render(<Field />);
-    const undoBtn = screen.getByText('Undo');
-    const redoBtn = screen.getByText('Redo');
-    fireEvent.click(undoBtn);
-    fireEvent.click(redoBtn);
-    expect(mockSdk.field.setValue).toHaveBeenCalledTimes(0); // editor is mocked
+    expect(screen.getByTestId('lottie-preview-field')).toContain({});
   });
 });
